Browser

+++ daw-m07/UF3/mvc_class_browser/App/views/browser/browseritemlist.php
@@ -6,26 +6,30 @@ if ($dh = opendir($ruta_user)) {
             //Creacio dels directoris amb les seves icones
             echo "<li>
             <img src='App/views/browser/icons/folder.png' alt='folder' width='15'>
-            <span><a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . "'>" . $file . "</a></span>
-            <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&delete=1' . "'>
-                <img src='App/views/browser/icons/fdelete.png' alt='delete' width='15'>
-            </a>
-            <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&zip=1' . "'>
-                <img src='App/views/browser/icons/zip.png' alt='zip' width='15'>
-            </a>
-            </li>";
+            <span><a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . "'>" . $file . "</a></span>";
+
+            if ($file != '.' && $file != '..') {
+                echo "
+                <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&delete=1' . "'>
+                    <img src='App/views/browser/icons/fdelete.png' alt='delete' width='15'>
+                </a>
+                <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&zip=1' . "'>
+                    <img src='App/views/browser/icons/zip.png' alt='zip' width='15'>
+                </a>";
+            }
+            echo "</li>";
         } else {
             //Creacio dels fitxers amb les seves icones
             echo "<li>
             <img src='App/views/browser/icons/file.png' alt='file' width='15'><span>" . $file . "</span>
-                <a href='?link=" . $link_user_sanitize . '\\' . $file . '&download=1' . "'>
+                <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&download=1' . "'>
                 <img src='App/views/browser/icons/down.png' alt='download' width='15'>
-                <a href='?link=" . $link_user_sanitize . '\\' . $file . '&view=1' . "'>
+                <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&view=1' . "'>
                 <img src='App/views/browser/icons/fview.png' alt='download' width='15'></a>
                 <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&zip=1' . "'>
                 <img src='App/views/browser/icons/zip.png' alt='zip' width='15'>
                 </a>
-                <a href='?link=" . $link_user_sanitize . '\\' . $file . '&delete=1' . "'>
+                <a href='index.php?action=browser&link=" . $link_user_sanitize . '\\' . $file . '&delete=1' . "'>
                 <img src='App/views/browser/icons/fdelete.png' alt='delete' width='15'>
                 </a>
                 </li>";
